Prepare for Websocket client return
